Add InsertPersonIntoOperation module
