+++ lib/operately/people.ex
@@ -1,9 +1,7 @@
 defmodule Operately.People do
   import Ecto.Query, warn: false
-  alias Ecto.Multi
-  alias Operately.Repo
 
-  alias Operately.Access
+  alias Operately.Repo
   alias Operately.People.{Person, Account}
 
   def list_people(company_id) do
@@ -30,49 +28,7 @@ defmodule Operately.People do
     if Operately.People.Account.valid_password?(account, password), do: account
   end
 
-  def insert_person(multi, callback) when is_function(callback, 1) do
-    multi
-    |> Multi.insert(:person, fn changes -> callback.(changes) end)
-    |> insert_person_access_group()
-    |> insert_membership_with_company_group()
-    |> insert_company_space_member()
-  end
-
-  defp insert_person_access_group(multi) do
-    multi
-    |> Multi.insert(:person_access_group, fn changes ->
-      Operately.Access.Group.changeset(%{person_id: changes.person.id})
-    end)
-    |> Multi.insert(:person_access_membership, fn changes ->
-      Operately.Access.GroupMembership.changeset(%{
-        group_id: changes.person_access_group.id,
-        person_id: changes.person.id,
-      })
-    end)
-  end
-
-  defp insert_membership_with_company_group(multi) do
-    multi
-    |> Multi.run(:company_access_group, fn _, %{person: person} ->
-      {:ok, Access.get_group!(company_id: person.company_id, tag: :standard)}
-    end)
-    |> Multi.insert(:company_access_membership, fn changes ->
-      Access.GroupMembership.changeset(%{
-        group_id: changes.company_access_group.id,
-        person_id: changes.person.id,
-      })
-    end)
-  end
-
-  defp insert_company_space_member(multi) do
-    multi
-    |> Multi.insert(:company_space_member, fn changes ->
-      Operately.Groups.Member.changeset(%{
-        group_id: changes.company_space.id,
-        person_id: changes.person.id,
-      })
-    end)
-  end
+  defdelegate insert_person(multi, callback), to: Operately.People.InsertPersonIntoOperation, as: :insert
 
   def create_person(attrs \\ %{}) do
     changeset = Person.changeset(%Person{}, attrs)
